+++ Cargo.toml
@@ -166,7 +166,7 @@ once_cell = "1.19.0"
 tempfile = "3.9.0"
 hex = "0.4.3"
 ethabi = "18.0.0"
-jsonrpsee = "0.22.3"
+jsonrpsee = "0.23.0"
 blake3 = "1.5.0"
 rand = "0.8.5"
 futures-util = "0.3.30"
