+++ crates/nox-tests/tests/chain_listener/mod.rs
@@ -16,17 +16,9 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-#![feature(try_blocks)]
 
-#[macro_use]
-extern crate fstrings;
-
-mod network {
-    pub use join::join_stream;
-
-    mod join;
-    mod loop_topology;
-    mod network_explore;
-}
-
-mod chain_listener;
+mod ccp_server;
+mod chain_server;
+mod r#const;
+mod tests;
+mod types;
